+++ Protos/src/thinclab/parsers/IPOMDPParser.java
@@ -42,7 +42,7 @@ public class IPOMDPParser extends ParseSPUDD {
 	public int frameID;
 	public String mostProbableAi = null;
 	public double mpAiProb = 0.99;
-	public HashMap<String, Double> aiDist = new HashMap<String, Double>();
+	public HashMap<String, Float> aiDist = new HashMap<String, Float>();
 	public List<ParseSPUDD> childFrames = new ArrayList<ParseSPUDD>();
 
 	private String domainPath = "";
@@ -190,7 +190,7 @@ public class IPOMDPParser extends ParseSPUDD {
 					}
 
 					else if (stream.sval.compareTo("most_probable_ai") == 0) {
-						this.aiDist.put(this.parseMostProbableAi(), this.parseMostProbableAiVal());
+						this.aiDist.put(this.parseMostProbableAi(), (float) this.parseMostProbableAiVal());
 						break;
 					}
